requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cloudinary==1.40.0
confusable-homoglyphs==3.3.1
dj-database-url==2.2.0
Django==5.0.6
django-debug-toolbar==4.4.2
django-registration==3.4
django-tinymce==4.0.0
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
